Add /pages, /page, /p subdomains for /chapters
Allows redirection to specific pages in a chapter. Works for /chapters, /chapters, and /c.

As with /chapters, /chapter, and /c, if the chapter doesn't exist, HnKRocks-Router will redirect to the Houseki no Kuni Manga title page on MangaDex.

If the chapter exists but the page doesn't, MangaDex just displays the first page of the chapter but keeps the "wrong" number in the address bar.

HnKRocks-Router makes no attempts to validate the page number beyond making sure it's an integer.

Increased hnkrocks-chapter-router version to 2.3.0