Add page selection to /latest, /newest, and /new
Adds the /pages, /page, /p subdomains for specific page redirection as was introduced for /chapters, /chapter, and /c in fe5b940.

The same rules apply, so HnKRocks-Router will not validate the provided page number beyond it being an integer and MangaDex will ignore out of bounds page numbers, defaulting to page 1.

Increased hnkrocks-chapter-router version to 2.4.0