Add /i for /pages, /page, and /p
This allows for rich embeds on Discord and other platforms. The specified page of the specified chapter will be displayed in an Embed which also contains the series title, the chapter's number, and the chapter uploader/scanlation group name.

The images are hosted in a Cloudflare R2 bucket. The HTML response uses the Twitter Card metadata format to allow for a large image display.

This endpoint works for /chapters, /chapter, and /c, as well as for /latest, /newest, and /new. For it to work, the page number has to be specified.

Increased hnkrocks-chapter-router version to 2.5.0