Add OEmbed endpoint for improved Embed rendering
This endpoint is used by automated systems such as Discord for requesting metadata regarding a link it is currently rendering as an embed.

The endpoint currently only adds "provider" information, which in Discord's case ends up getting rendered as a small text at the top of the Embed, reading some variation of "HnK.Rocks - [Feature Description]". This provider information also links to the /about endpoint.

Additional OEmbed data may be added in the future.

Increased hnkrocks-chapter-router version to 2.7.0