Add redirects for various interviews
Added the new /interviews/... section which will contain several interviews conducted with Haruko Ichikawa as well as other people associated with the works of Ichikawa. It uses a language parameter to allow for future addition of Japanese-only interviews (or other languages).

- `/interviews/en/eureka-2010` and `/interviews/en/eureka` redirect to an interview with Eureka from 2010
- `/interviews/en/kono-manga-2014-pt-1` and `/interviews/en/2014-pt-1` redirects to part 1 of an interview with Kono Manga ga Sugoi from 2014
- `/interviews/en/kono-manga-2014-pt-2` and `/interviews/en/2014-pt-2` redirects to part 2 of the above-mentioned interview
- `/interviews/en/entertainment-week-2016`, `/interviews/en/entertainment-week`, and `/interviews/en/entertainment-2016` all redirect to an interview with Entertainment Week from 2016

Moved the 2015 interview with Current Obsession from `/other/current-obsession` to `/interviews/en/current-obsession-2015`, `/interviews/en/current-obsession`, and `/interviews/en/co-2015`. A shim for the previous routing was added as well.

Increased hnkrocks-chapter-router version to 2.14.0