Add /c/X/p/Y/i for Minimalist and Colour versions
This adds direct page embedding support for the Minimalist Colour and Coloured versions of Houseki no Kuni.

Chapter links and chapter page links are NOT supported as they are unavailable.

- `/minimalist/c/1/p/1/i` will result in a rich embed for chapter 1, page 1 of the Minimalist Colour version of Houseki no Kuni (all of `chapters`, `chapter`, and `c` as well as all of `pages`, `page`, and `p` are supported in the URL)

- `/coloured/c/1/p/1/i` will result in a rich embed for chapter 1, page 1 of the Coloured version of Houseki no Kuni (same aliases as for the Minimalist version are supported)

As a side effect, the direct page response generation was refactored into a utility function to be reused by all direct page embedding handlers.

Increased hnkrocks-chapter-router version to 2.27.0